refactor(app): extract status-view action policy
